feat: 가구 기능 구현

+++ src/main/java/com/saram/jellylog/furniture/dto/request/FurnitureUpdateRequest.java
@@ -3,7 +3,7 @@ package com.saram.jellylog.furniture.dto.request;
 public record FurnitureUpdateRequest(
     String furnitureName,
     String furnitureImage,
-    Long furniturePrice,
-    Long furnitureGroup
+    Integer furniturePrice,
+    Integer furnitureGroup
 ) {
 }
